backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
PyJWT==2.9.0
bcrypt==4.2.0
pyodbc==5.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
